Add the Queue class
A Queue object stores a list of event in order, and it has two states
"ready" and "not ready".  When a Queue object is in ready state, it
means that the events it stores are ready for upload.  You may configure
a Queue object for when a Queue should be ready.

Queue objects are proactive in the sense that you do not periodically
check whether a Queue is ready.  Instead, a Queue will call you when it
is ready (uploadRequester.requestUpload).