Add Uploader
An Uploader object accepts list of objects (called a "batch") and
uploads batches in order and in the background.

An Uploader object implements exponential backoff to handle upload
failures (we might tune the parameters once we have more data).

If a batch has been rejected by the server more than a certain times,
the Uploader object will drop that batch.  A batch could be rejected for
a number of reasons, like data are corrupted or the data schema is
incorrect.  Whatever the reason is, the Uploader object drops that batch
so that it could make some progress (uploading batches behind this one).

NOTE: At the moment, the Uploader might keep working even when all
activities are not visible (i.e., stopped or destroyed).  This will be
fixed later.